Add Info.plist and command line arguments
Adds command-line arguments for the display name and identifier for the new version of the framework that is being compiled. Also adds a new version display name argument in the Info.plist, since the bundle version is being used as the identifier.

Currently, the version display name and identifier are being taken from the Info.plist and the command line arguments are not used, but eventually the command line arguments for the identifier and display name would be used to override an outdated Info.plist.