Encode patches in RenderNodes and RenderIndex
For each RenderNode, uses the encoder to store the previous version’s RenderNode and uses difference(from:at:) to determine the difference between each RenderNode and its previous version. These differences are encoded in the node’s ‘versions’ property. The same is done for each RenderIndex.

More info:

In Convert, I copy over the contents of the previous archive to the temp directory that the new archive will be written into. Currently, I only copy over the index and data directories, but for instance for more complex frameworks, whole archives would probably need to be copied over to account for cases where images/assets are deleted in new versions.

When each new RenderNode is about to written by the JSONEncoderRenderNodeWriter, I decode the Render JSON file that is already there, and put it in the new previousRenderNode property in the userInfo of the encoder. Whatever previous render node was there will be overwritten.

Then, when the new RenderNode is being encoded, I take the previous render node from the encoder and call the difference(from:at:) method from Diffable on it (see below subsection for detail), passing in the current node for it to diff against. It will return an array of JSONPatchOperations—the same ones that are used for the Objective-C variants.

The [JSONPatchOperation] is put into a VersionPatch, which holds the patches required to convert the RenderNode back into the previous version, as well as the ArchiveVersion (name and identifier) of the version the patches will switch the RenderNode back to.